Refactor and move api/types/authentication.go to lib/services.
